[refactor] 코드정리
