+++ libs/components-seller/src/lib/SellerMainFeatureSection.tsx
@@ -1,117 +1,10 @@
-import {
-  Box,
-  BoxProps,
-  Container,
-  Stack,
-  Text,
-  useColorModeValue,
-} from '@chakra-ui/react';
+import { Stack, Text, useColorModeValue } from '@chakra-ui/react';
 import {
   featureSectionBody,
   SectionData,
   sellerMainSectionText,
 } from '@project-lc/components-constants/sellerMainText';
-import { motion } from 'framer-motion';
-import React from 'react';
-
-/** 크크쇼 판매자센터 메인 섹션 타이틀 옆 애니메이션 막대 */
-export function FadeInBar(): JSX.Element {
-  return (
-    <Box
-      position={{ base: 'relative', xl: 'absolute' }}
-      left={{ base: 0, xl: '-80px' }}
-      top={{ base: '-0.5rem', xl: '1.5rem' }}
-      w="80px"
-      height="5px"
-      display="inline-block"
-    >
-      <motion.div
-        style={{
-          position: 'absolute',
-          top: 0,
-          bottom: 0,
-          left: 0,
-          right: 0,
-          background: 'var(--chakra-colors-blue-500)',
-        }}
-        initial={{
-          right: '100%',
-          opacity: 0,
-        }}
-        whileInView={{
-          right: 0,
-          opacity: 1,
-        }}
-        transition={{ duration: 1, ease: 'easeInOut' }}
-        viewport={{ once: true }}
-      />
-    </Box>
-  );
-}
-
-/** 크크쇼 판매자센터 메인페이지 섹션 공용 레이아웃 컴포넌트 */
-export function SellerMainSectionContainer({
-  children,
-  sectionData,
-  bgProps,
-  hasGrayBg = false,
-}: {
-  hasGrayBg?: boolean;
-  children?: React.ReactNode;
-  sectionData: SectionData;
-  bgProps?: BoxProps;
-}): JSX.Element {
-  const { title, desc } = sectionData;
-  const grayBg = useColorModeValue('gray.50', 'gray.700');
-
-  return (
-    // 배경 컨테이너
-    <Box
-      py={{ base: 10, sm: 14, xl: 20 }}
-      position="relative"
-      overflow="hidden"
-      bg={hasGrayBg ? grayBg : undefined}
-      {...bgProps}
-    >
-      {/* 실제 컨텐츠 영역(최대너비 1200px임, breakpoint는 xl(1280px) 사용) */}
-      <Container maxW="container.xl" position="relative">
-        <Stack spacing={{ base: 6, sm: 8 }} position="relative">
-          <Box
-            pl={{ base: 0, md: '1.5rem' }}
-            textAlign={{ base: 'center', md: 'left' }}
-            position="relative"
-          >
-            <FadeInBar />
-
-            <Stack wordBreak="keep-all">
-              {title && (
-                <Text
-                  fontFamily="Gmarket Sans"
-                  fontWeight="bold"
-                  lineHeight="shorter"
-                  whiteSpace="pre-line"
-                  fontSize={{ base: '2xl', sm: '3xl' }}
-                >
-                  {title}
-                </Text>
-              )}
-              {desc && (
-                <Text
-                  whiteSpace={{ base: 'pre-line', md: 'normal' }}
-                  wordBreak="keep-all"
-                  fontSize={{ base: 'sm', sm: 'lg' }}
-                >
-                  {desc}
-                </Text>
-              )}
-            </Stack>
-          </Box>
-          {children}
-        </Stack>
-      </Container>
-    </Box>
-  );
-}
+import { SellerMainSectionContainer } from '@project-lc/components-layout/SellerMainSectionContainer';
 
 /** 크크쇼 판매자센터 특징 아이템 */
 export function SellerMainFeatureItem({ item }: { item: SectionData }): JSX.Element {
